requirements.txt
google-auth==2.16.0
google-api-python-client==2.73.0